+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
-## [5.8.0] - 2018-08-14
+## [6.0.0] - 2018-08-14
 ### Added
 - Bow enhancement.
+- Conditional operator.
+- Range operator.
+- Exclusion operator.
+- Possibility operator.
 ### Fixed
 - Fixed an error that prevents gui from opening.
 ### Changed
 - Renamed `enableFireworkDamage: false` to `enablePreventFireworkDamage: true`.
-### Removed
 
 ## [5.7.0] - 2018-08-13
 ### Added